+++ src/Commands/BackupCommands.php
@@ -22,10 +22,10 @@ class BackupCommands extends Command
 
         $this->info('Starting backup');
 
-
-        /**
-         * Check if the user has access to read and write files, does the user have permission?
-         */
+        try {
+            /**
+             * Check if the user has access to read and write files, does the user have permission?
+             */
 
         foreach ($this->getTargetFileSystems() as $fileSystem) {
             $this->backupHelper->checkIfUserHasPermissions($fileSystem);
@@ -50,13 +50,13 @@ class BackupCommands extends Command
 
         $compress = config('db-manager.output.compress');
 
-        if(!isset($compress) || (isset($compress) && !is_bool($compress))) {
+        if (!isset($compress) || (isset($compress) && !is_bool($compress))) {
             $compress = true;
         }
 
         $this->info("Compress output files: " . ($compress ? "True" : "False"));
 
-        if($compress) {
+        if ($compress) {
 
             $backupZipFile = $this->createZip($files);
 
@@ -100,10 +100,18 @@ class BackupCommands extends Command
             }
         }
 
+        } catch (\Exception $e) {
+            $this->warn('An Error occurred');
+            $this->warn("Code: " . $e->getCode());
+            $this->warn("Message: \n". $e->getMessage());
+            $this->warn("Starck Trace: \n" . $e->getTraceAsString());
+        }
 
         /**
          * Delete the temporary files
          */
+
+        $this->info("Cleaning up temporary files");
 
         if(!$this->backupHelper->cleanUpTemporaryFiles()) {
             $this->warn('Unable to remove temporary files, may need to be manually removed');
@@ -112,198 +120,6 @@ class BackupCommands extends Command
         $this->info('Backup successfully completed');
 
         return true;
-    }
-
-    /**
-     * Provides a list of all the database tables to be backed up
-     *
-     * @return array
-     * @throws \Exception
-     */
-    protected function getAllTablesToBeBackedUp()
-    {
-        $files = [];
-
-        $files[] = ['realFile' => $this->getDatabaseDump($files), 'fileInZip' => 'dump.sql'];
-
-        return $files;
-    }
-
-    /**
-     * Creates a zip archive from the file supplied
-     *
-     * @param $files
-     * @return string
-     */
-    protected function createZip($files)
-    {
-        $this->comment('Start zipping '.count($files).' files...');
-
-        $tempZipFile = $this->backupHelper->getTemporaryFileDir() . $this->backupHelper->getTemporaryFileName();
-
-        $zip = new ZipArchive();
-        $zip->open($tempZipFile, ZipArchive::CREATE);
-
-        foreach ($files as $file) {
-            if (file_exists($file['realFile'])) {
-                $zip->addFile($file['realFile'], $file['fileInZip']);
-            }
-        }
-
-        $zip->close();
-
-        chmod($tempZipFile, 0777);
-
-        $this->comment('Zip created!');
-
-        return $tempZipFile;
-    }
-
-
-    /**
-     * Get the filesystem in use from the config
-     *
-     * @return array|mixed
-     */
-    protected function getTargetFileSystems()
-    {
-        $fileSystems = config('db-manager.output.filesystem');
-
-        if (is_array($fileSystems)) {
-            return $fileSystems;
-        }
-
-        $arrayString = $this->is_array_string($fileSystems);
-
-        if (is_array($arrayString)) {
-            return $arrayString;
-        }
-
-        return [$fileSystems];
-    }
-
-    protected function is_array_string($string)
-    {
-        if(strlen($string) <= 2) {
-            return false;
-        }
-
-        if( $string[0] != '[' ||
-            $string[strlen($string)-1] != ']') {
-             return false;
-        }
-
-        try {
-            $arr = explode(',', substr($string, 1, strlen($string) - 2));
-            if(count($arr) <= 0) {
-                return false;
-            }
-        } catch (\Exception $e) {
-            return false;
-        }
-
-        $arr = array_map(function($obj) {
-            $bad = ['\'', '"'];
-            return trim(str_replace($bad, "", $obj));
-        }, $arr);
-
-        return $arr;
-    }
-
-    /**
-     * Get the backup location destination full path
-     *
-     * @return string
-     * @throws \Exception
-     */
-    protected function getBackupDestinationFileName()
-    {
-        $backupDirectory = config('db-manager.output.location');
-        $backupFilename = $this->getPrefix().$this->getFilename().$this->getSuffix().$this->backupHelper->getOutputFileType();
-
-        return $backupDirectory.'/'.$backupFilename;
-    }
-
-    /**
-     * Get the prefix of the output filename from the config
-     *
-     * @return array|bool|string
-     */
-    public function getPrefix()
-    {
-        if ($this->option('prefix') != '') {
-            return $this->option('prefix');
-        }
-
-        $prefix = $this->getFixSpecialType(config('db-manager.output.prefix'));
-
-        return $prefix;
-    }
-
-    /**
-     * Get the filename from the config for the output file name
-     *
-     * @return mixed
-     * @throws \Exception
-     */
-    public function getFilename()
-    {
-        if (empty(config('db-manager.output.filename'))) {
-            throw new \Exception('Filename not set in config');
-        }
-
-        return config('db-manager.output.filename');
-    }
-
-    /**
-     * Get the suffix of the output file
-     *
-     * @return array|bool|string
-     *
-     */
-    public function getSuffix()
-    {
-        if ($this->option('suffix') != '') {
-            return $this->option('suffix');
-        }
-
-        $suffix = $this->getFixSpecialType(config('db-manager.output.suffix'));
-
-        return $suffix;
-    }
-
-    /**
-     * A method that returns a special variable for the file name supplied
-     *
-     * @param $fix
-     * @return bool|string
-     */
-    public function getFixSpecialType($fix)
-    {
-        if($fix == 'datetime') {
-            return date('YmdHis');
-        } else {
-            return $fix;
-        }
-    }
-
-    /**
-     * Copy a supplied file to the filesystem passed through in the destination specified in the config
-     *
-     * @param $file
-     * @param $fileSystem
-     */
-    public function copyFileToFileSystem($file, $fileSystem)
-    {
-        $this->comment('Start uploading backup to '.$fileSystem.'-filesystem...');
-
-        $backupFilename = $this->getBackupDestinationFileName();
-
-        if($this->backupHelper->copyFileToFileSystem($file, $fileSystem, $backupFilename)) {
-            $this->comment('Backup stored on '.$fileSystem.'-filesystem in file "'.$backupFilename.'"');
-        } else {
-            $this->warn('Unable to send file "'.$backupFilename.'" to '.$fileSystem.'-filesystem');
-        }
     }
 
     /**
@@ -321,59 +137,6 @@ class BackupCommands extends Command
             ['keeplastonly', null, InputOption::VALUE_REQUIRED, 'Keep the last backup or delete all previous backups (true/false)'],
             ['compress', null, InputOption::VALUE_REQUIRED, 'Compress the output file to .zip (true/false)'],
         ];
-    }
-
-    /**
-     * Get the file from the database dump
-     *
-     * @return mixed
-     * @throws \Exception
-     */
-    protected function getDatabaseDump()
-    {
-        $filesToBeBackedUp = $this->backupHelper->getFilesToBeBackedUp($this);
-
-        if (count($filesToBeBackedUp) != 1) {
-            throw new \Exception('could not backup db');
-        }
-
-        $this->comment('Database dumped');
-
-        return $filesToBeBackedUp[0];
-    }
-
-    /**
-     * Delete all previous backups from the selected filesystem
-     *
-     * @param $fileSystem
-     * @return bool
-     */
-    public function deletePreviousBackups($fileSystem)
-    {
-        do {
-            $name = $this->ask('Are you sure you want to remove all previous backups from the filesystem '.$fileSystem.'? [y/n]');
-            if(strcasecmp ($name, "y") != 0 && strcasecmp ($name, "n") != 0) {
-                $this->info("Invalid response, type 'y' for yes or 'n' for no. Let's try again. This is serious, if it wasn't, I would've use the confirm method to write this");
-            }
-
-        } while(strcasecmp ($name, "y") != 0 && strcasecmp ($name, "n") != 0);
-
-
-        if (strcasecmp ($name, "y") != 0) {
-            $this->info("Skipped deleting all previous backups");
-            return true;
-        }
-
-        $this->info("Deleting all previous backups");
-
-        $deletedFiles = $this->backupHelper->deleteTargetDirectoryFiles($fileSystem);
-
-        foreach($deletedFiles as $file) {
-            $this->info("Deleted backup file " . $file);
-        }
-
-        return true;
-
     }
 
 }