+++ src/DBManager.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace JulianPitt\DBManager;
+
+use Illuminate\Support\Facades\Facade;
+
+class DBManager extends Facade {
+
+    protected static function getFacadeAccessor() {
+        return '\JulianPitt\DBManager\DBManagerClass';
+    }
+
+}