+++ src/DBManagerClass.php
@@ -2,15 +2,25 @@
 
 namespace JulianPitt\DBManager;
 
-class DBManager
+use JulianPitt\DBManager\Commands\BackupCommands;
+use JulianPitt\DBManager\Commands\RestoreCommands;
+use JulianPitt\DBManager\Helpers\BackupHelper;
+
+class DBManagerClass
 {
     /**
-     * Create a new DBDataManager Instance
+     * Create a new DBManagerClass Instance
      */
+
     public function __construct()
     {
         // constructor body
-        new Commands\BackupCommand();
+        $this->backupClass = new BackupHelper();
+    }
+
+    public function hasPermission()
+    {
+        return $this->backupClass->checkIfUserHasPermissions('local');
     }
 
 }